+++ packages/local_auth/local_auth/CHANGELOG.md
@@ -1,7 +1,40 @@
 ## 2.0.0
 
-* Migrated plugin to federated architecture.
+* Migrates plugin to federated architecture.
 * Adds OS version support information to README.
+* BREAKING CHANGE: Deprecated method `authenticateWithBiometrics` has been removed
+  Use `authenticate` instead.
+* BREAKING CHANGE: Parameters of `authenticate` have been changed.
+  
+  Example:
+```dart  
+// Old way of calling `authenticate`.
+Future<bool> authenticate(
+  localizedReason: 'localized reason',
+  useErrorDialogs: true,
+  stickyAuth: false,
+  androidAuthStrings: const AndroidAuthMessages(),
+  iOSAuthStrings: const IOSAuthMessages(),
+  sensitiveTransaction: true,
+  biometricOnly: false,
+);
+// New way of calling `authenticate`.
+Future<bool> authenticate(
+  localizedReason: 'localized reason',
+  authMessages: const <AuthMessages>[
+    IOSAuthMessages(),
+    AndroidAuthMessages()
+  ],
+  options: const AuthenticationOptions(
+    useErrorDialogs: true,
+    stickyAuth: false,
+    sensitiveTransaction: true,
+    biometricOnly: false,
+  ),
+);
+```
+
+
 
 ## 1.1.11